+++ src/core/env.ts
@@ -2,6 +2,7 @@ import { environment } from "./config";
 
 type DOMState = "soft-loading" | "hard-loading" | "idling" | "page-loading" | "page-loading-complete";
 type NetworkType = "4g" | "3g" | "2g" | "slow-2g";
+type Browser = "chrome" | "safari" | "edge" | "chromium-edge" | "ie" | "firefox" | "unknown" | "opera";
 
 class Env {
     public isDebug: boolean;
@@ -11,6 +12,7 @@ class Env {
     public isProduciton: boolean;
     public domState: DOMState;
     public dataSaver: boolean;
+    public browser: Browser;
 
     private _tickets: Array<string>;
 
@@ -23,6 +25,8 @@ class Env {
         this.isDebug = !this.isProduciton;
         this.domState = "hard-loading";
         this.dataSaver = false;
+        this.browser = "unknown";
+        this.setBrowser();
 
         this._tickets = [];
 
@@ -138,8 +142,54 @@ class Env {
         this.domState = newState;
         document.documentElement.setAttribute("state", this.domState);
     }
+
+    private setBrowser() {
+        // @ts-ignore
+        var isOpera = (!!window.opr && !!opr.addons) || !!window.opera || navigator.userAgent.indexOf(" OPR/") >= 0;
+
+        // @ts-ignore
+        var isFirefox = typeof InstallTrigger !== "undefined";
+
+        var isSafari =
+            // @ts-ignore
+            /constructor/i.test(window.HTMLElement) ||
+            (function(p) {
+                return p.toString() === "[object SafariRemoteNotification]";
+                // @ts-ignore
+            })(!window["safari"] || (typeof safari !== "undefined" && safari.pushNotification));
+
+        // @ts-ignore
+        var isIE = /*@cc_on!@*/ false || !!document.documentMode;
+
+        var isEdge = !isIE && !!window.StyleMedia;
+
+        // @ts-ignore
+        var isChrome = !!window.chrome && (!!window.chrome.webstore || !!window.chrome.runtime);
+
+        var isEdgeChromium = isChrome && navigator.userAgent.indexOf("Edg") != -1;
+
+        if (isOpera) {
+            this.browser = "opera";
+        } else if (isFirefox) {
+            this.browser = "firefox";
+        } else if (isSafari) {
+            this.browser = "safari";
+        } else if (isIE) {
+            this.browser = "ie";
+        } else if (isEdge) {
+            this.browser = "edge";
+        } else if (isChrome) {
+            this.browser = "chrome";
+        } else if (isEdgeChromium) {
+            this.browser = "chromium-edge";
+        } else {
+            this.browser = "unknown";
+        }
+        document.documentElement.classList.add(this.browser);
+    }
 }
 export const env: Env = new Env();
 export const debug: boolean = env.isDebug;
 export const uuid: Function = env.uuid;
 export const dataSaver: boolean = env.dataSaver;
+export const browser: Browser = env.browser;